+++ CHANGELOG.md
@@ -2,7 +2,7 @@ Version 7.0.0-dev
 - BC break: used interfaces for data provider and skip checker in TestCase
 - BC break: used typed class constants
 - BC break: made some methods in TAssertions final
-- BC break: it is no longer possible to pass null to parameter $resultsFormatter in Tester's constructor
+- BC break: it is no longer possible to pass null to parameters $resultsFormatter and $testSuitesFinder in Tester's constructor
 - BC break: made Tester::$testSuitesFinder readonly
 - BC break: added method IResultsFormatter::setOutputFileName
 - BC break: added methods ITesterExtension::getEventsBeforeTestCase and getEventsAfterTestCase